ServidorIA/requirements.txt
flask
opencv-python-headless
ultralytics
numpy
datasets
pillow
tqdm
pyyaml
